+++ userspace/lib/graphics.h
@@ -72,7 +72,8 @@ void draw_line(gfx_context_t * ctx, int32_t x0, int32_t x1, int32_t y0, int32_t 
 void draw_line_thick(gfx_context_t * ctx, int32_t x0, int32_t x1, int32_t y0, int32_t y1, uint32_t color, char thickness);
 void draw_fill(gfx_context_t * ctx, uint32_t color);
 
-void draw_sprite_scaled(gfx_context_t * ctx, sprite_t * sprite, uint16_t x, uint16_t y, uint16_t width, uint16_t height);
+void draw_sprite_scaled(gfx_context_t * ctx, sprite_t * sprite, int32_t x, int32_t y, uint16_t width, uint16_t height);
+
 void context_to_png(FILE * file, gfx_context_t * ctx);
 
 #endif
